refactor(documentation): propSwitcher on change event
